% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cengiz_df.R
\docType{data}
\name{Cengiz_df}
\alias{Cengiz_df}
\title{Minimum wage dataset}
\format{
A data frame with 214812 rows and 9 variables:
\describe{
\item{wagebins}{Wage bin}
\item{statenum}{State ID}
\item{year}{Year}
\item{quarterdate}{Quarter}
\item{overallcountpc}{Employment per capita in wage bin}
\item{treated_quarter}{Was state treated that quarter}
\item{treated_year}{Was state treated last 4 consecutive quarters}
\item{MW}{Actual minimum wage}
\item{population}{State population}
}
}
\source{
Roth and Sant'Anna (2023)
}
\usage{
Cengiz_df
}
\description{
A dataset containing number of people at each wage level in 25c bins, from
2007 to 2015.  This is a subset of the dataset used in Roth and Sant'Anna
(2023) that was in turn is derived from Cengiz et al. (2019). See those
papers for additional descriptions.
}
\references{
\cite{Cengiz, Doruk, Dube, Arindrajit, Lindner, Attila and Zipperer, Ben (2019).
"The Effect of Minimum Wages on Low-Wage Jobs."
The Quarterly Journal of Economics, vol. 134 (3), pp. 1405–1454, \doi{10.1093/qje/qjz014}}

\cite{Roth, Jonathan and Sant'Anna, Pedro H. C. (2023),
"When is Parallel Trends Sensitive to Functional Form?"
Econometrica, vol. 91 (2), pp. 737–747, \doi{10.3982/ECTA19402}}
}
\keyword{datasets}
