% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{didFF-package}
\alias{didFF-package}
\title{Difference-in-Differences and Functional Form}
\description{
The \code{didFF} R package assesses when the validity of difference-in-differences and related estimators depends on functional form, based on the theoretical results in \href{https://doi.org/10.3982/ECTA19402}{Roth and Sant’Anna (2023)}.
The package provides a test for whether parallel trends is insensitive to functional form by estimating the implied density of potential outcomes under the null and checking if it is significantly below zero at some point.
It also provides a distributional DiD estimators based on a distributional parallel trends assumption.
}
\references{
\cite{Roth, Jonathan and Sant'Anna, Pedro H. C. (2023),
"When is Parallel Trends Sensitive to Functional Form?"
Econometrica, vol. 91 (2), pp. 737–747, \doi{10.3982/ECTA19402}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psantanna.com/didFF/}
  \item \url{https://github.com/pedrohcgs/didFF/}
  \item \url{http://psantanna.com/didFF/}
}

}
\author{
\strong{Maintainer}: Pedro H. C. Sant'Anna \email{pedrohcgs@gmail.com}

Authors:
\itemize{
  \item Jonathan Roth \email{jonathndr@gmail.com}
}

}
\keyword{internal}
