% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distDD.R
\name{distDD}
\alias{distDD}
\title{Distributional treatment effect}
\usage{
distDD(...)
}
\value{
A list object containing: The average treatment effects; a
table with the estimated distributional TE, their SE, and bin levels.
}
\description{
Wrapper for \code{didFF(..., distDD=TRUE)}, which computes
the distributional treatment effects. This estimates the actual
distribution of \code{Y(1)} minus the implied distribution for \code{Y(0)} in
each outcome bin (for the treated group).
}
