% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctionalForm.R
\name{TestFunctionalForm}
\alias{TestFunctionalForm}
\title{Test if Parallel trends assumption is sensitive to functional form}
\usage{
TestFunctionalForm(
  DF = NULL,
  idvar = "id",
  yvar = "output",
  tvar = "time",
  treatmentvar = "treatment group",
  weight = NULL,
  scale = NULL,
  nbins = NULL,
  nboots = 1000,
  seed = 0,
  starting_year = NULL,
  ending_year = NULL,
  minbin = NULL,
  maxbin = NULL
)
}
\arguments{
\item{DF}{data.frame object}

\item{idvar}{Name variable indicating group or id in DF}

\item{yvar}{Name of Outcome variable in DF}

\item{tvar}{Name Time variable in DF}

\item{treatmentvar}{Name of Binary variable in df indicating treatment/control group}

\item{weight}{Name weighting variable. Default is NULL}

\item{scale}{Name scale variable. Default is NULL}

\item{nbins}{A scalar indicating the Number of bins for the support of Outcome. Default nbins=100}

\item{nboots}{Number of bootstrap samples for iDensityTest. Default nboots=1000}

\item{seed}{Starting seed for iDensityTest. Default is random.}

\item{starting_year}{The Baseline period for DID. Default min(tvar)}

\item{ending_year}{The follow up period for DID. Default max(tvar)}

\item{minbin}{Minimun outcome-bin for density estimation. Default minbin=NULL}

\item{maxbin}{Maximun outcome-bin for density estimation. Default maxbin=NULL}
}
\value{
A plot of the implied density under the null and pval for H0= Implied Density>0
}
\description{
Test if Parallel trends assumption is sensitive to functional form
}
