% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctionalForm.R
\name{TestFunctionalForm}
\alias{TestFunctionalForm}
\title{Test if Parallel trends assumption is sensitive to functional form}
\usage{
TestFunctionalForm(
  DF = NULL,
  idvar = "id",
  yvar = "output",
  tvar = "time",
  treatmentvar = "treatment group",
  nbins = 100,
  nboot = 1000,
  starting_year = NULL,
  ending_year = NULL
)
}
\arguments{
\item{DF}{data.frame object}

\item{idvar}{Name variable indicating group or id in DF}

\item{yvar}{Name of Outcome variable in DF}

\item{tvar}{Name Time variable in DF}

\item{treatmentvar}{Name of Binary variable in df indicating treatment/control group}

\item{nbins}{A scalar indicating the Number of bins for the support of Outcome. Default nbins=100}

\item{nboot}{Number of bootstrap for iDensityTest. Default nboot=1000}

\item{starting_year}{The Baseline period for DID. Default min(tvar)}

\item{ending_year}{The follow up period for DID. Default max(tvar)}
}
\value{
A plot of the implied density under the null
}
\description{
Test if Parallel trends assumption is sensitive to functional form
}
\examples{
set.seed(99)
N=10000
treat<-rbinom(N,1,0.25)
wage<-rlnorm(N, 13, 2)
id<-as.character(seq(1:N))
t0<-replicate(N,2010)
t1<-replicate(N,2020)
y0<-wage+rnorm(N,0,100)
y1<- y0 + 20*treat + rnorm(N,0,100)
panel_id=c(paste(id,t0,sep="_"),paste(id,t1,sep="_"))
year<-c(t0,t1)
y<-c(y0,y1)
D<-c(treat,treat)
treated<-c(replicate(N,0),treat)
ID<-c(id,id)
df<-data.frame(panel_id,ID,year,y,D,treated)
df<-df\%>\%dplyr::filter((0 <= y) & (y <= 500000))
TestFunctionalForm(df,"ID","y","year","D",100,1000)
}
